+++ lib/_core/constants/http.dart
@@ -3,7 +3,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:logger/logger.dart';
 
 // 192.168.219.111
-final baseURL = "http://192.168.0.101:8080";
+final baseURL = "http://192.168.219.110:8080";
 
 final dio = Dio(
   BaseOptions(
