+++ lib/ui/holder/codi/codi_insert_page/codi_insert_viewmodel.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
-
 import 'package:final_project_team02/data/dtos/codi_req.dart';
 import 'package:final_project_team02/data/dtos/respons_dto.dart';
 import 'package:final_project_team02/data/repositoreis/codi_repo.dart';
 import 'package:final_project_team02/data/session_data/session_data.dart';
 import 'package:final_project_team02/main.dart';
+import 'package:final_project_team02/ui/holder/codi/codi_insert_page/codi_insert_data/codi_insert.dart';
 import 'package:final_project_team02/ui/holder/home/home_data/popular_codi_photos.dart';
 import 'package:final_project_team02/ui/holder/home/home_viewmodel.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator/creator_data/creator_codi_list.dart';
@@ -14,87 +14,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:path/path.dart' as path;
-
-class CodiInsertModel {
-  final List<XFile> images;
-  final String prevImg;
-
-  final String? fileName; // 파일이름
-  final String? fileExtension; // 확장자
-  final bool? isMainPhoto;
-
-  final int? topItemId;
-  final int? topBrandId;
-  final String? topPhotoPath;
-  final int? topImageId;
-  final String? topItemName;
-
-  final int? bottomItemId;
-  final int? bottomBrandId;
-  final String? bottomPhotoPath;
-  final int? bottomImageId;
-  final String? bottomItemName;
-
-  final String comment;
-
-  const CodiInsertModel({
-    required this.images,
-    required this.prevImg,
-    this.fileName,
-    this.fileExtension,
-    this.isMainPhoto,
-    this.topItemId,
-    this.topBrandId,
-    this.topPhotoPath,
-    this.topImageId,
-    this.topItemName,
-    this.bottomItemId,
-    this.bottomBrandId,
-    this.bottomPhotoPath,
-    this.bottomImageId,
-    this.bottomItemName,
-    required this.comment,
-  });
-
-  CodiInsertModel copyWith({
-    List<XFile>? images,
-    String? prevImg,
-    String? fileName,
-    String? fileExtension,
-    bool? isMainPhoto,
-    int? topItemId,
-    int? topBrandId,
-    String? topPhotoPath,
-    int? topImageId,
-    String? topItemName,
-    int? bottomItemId,
-    int? bottomBrandId,
-    String? bottomPhotoPath,
-    int? bottomImageId,
-    String? bottomItemName,
-    String? comment,
-  }) {
-    return CodiInsertModel(
-      images: images ?? this.images,
-      prevImg: prevImg ?? this.prevImg,
-      fileName: fileName ?? this.fileName,
-      fileExtension: fileExtension ?? this.fileExtension,
-      isMainPhoto: isMainPhoto ?? this.isMainPhoto,
-      topItemId: topItemId ?? this.topItemId,
-      topBrandId: topBrandId ?? this.topBrandId,
-      topPhotoPath: topPhotoPath ?? this.topPhotoPath,
-      topImageId: topImageId ?? this.topImageId,
-      topItemName: topItemName ?? this.topItemName,
-      bottomItemId: bottomItemId ?? this.bottomItemId,
-      bottomBrandId: bottomBrandId ?? this.bottomBrandId,
-      bottomPhotoPath: bottomPhotoPath ?? this.bottomPhotoPath,
-      bottomImageId: bottomImageId ?? this.bottomImageId,
-      bottomItemName: bottomItemName ?? this.bottomItemName,
-      comment: comment ?? this.comment,
-    );
-  }
-}
+import 'package:path/path.dart';
 
 class CodiInsertViewModel extends StateNotifier<CodiInsertModel> {
   Ref ref;
@@ -128,8 +48,13 @@ class CodiInsertViewModel extends StateNotifier<CodiInsertModel> {
     }
   }
 
-  Future<void> pickAndAddImageFromBase64(String photoPath, String itemName,
-      int itemId, int brandId, String category) async {
+  Future<void> pickAndAddItemImage(
+    String photoPath,
+    String itemName,
+    int itemId,
+    int brandId,
+    String category,
+  ) async {
     if (category == 'top') {
       state = state.copyWith(
         topBrandId: brandId,
@@ -158,14 +83,14 @@ class CodiInsertViewModel extends StateNotifier<CodiInsertModel> {
     }
   }
 
-  Future<void> pickAndAddImage() async {
+  Future<void> pickAndAddCodiImage() async {
     final XFile? image = await picker.pickImage(source: ImageSource.gallery);
     if (image != null) {
       String base64String = await convertToBase64(image);
 
       // 파일 경로에서 파일 이름과 확장자 추출
-      String fileName = path.basename(image.path);
-      String fileExtension = path.extension(image.path);
+      String fileName =basename(image.path);
+      String fileExtension = extension(image.path);
 
       await cacheManager.putFile(
         image.path,
