+++ lib/ui/holder/codi/codi_insert_page/components/codi_insert_page_body.dart
@@ -1,5 +1,6 @@
 import 'package:final_project_team02/data/dtos/codi_req.dart';
 import 'package:final_project_team02/data/session_data/session_data.dart';
+import 'package:final_project_team02/ui/holder/codi/codi_insert_page/codi_insert_data/codi_insert.dart';
 import 'package:final_project_team02/ui/holder/codi/codi_insert_page/codi_insert_viewmodel.dart';
 import 'package:final_project_team02/ui/holder/codi/codi_insert_page/components/codi_insert_styling_comment.dart';
 import 'package:final_project_team02/ui/holder/codi/codi_insert_page/components/codi_insert_sub_title.dart';
