+++ lib/ui/holder/codi/codi_item_insert_page/components/codi_item_brand_tab_view.dart
@@ -35,7 +35,7 @@ class CodiItemBrandTabView extends ConsumerWidget {
         final brand = model.brandList[selectedIndex];
         return GestureDetector(
           onTap: () {
-            ref.read(codiInsertProvider.notifier).pickAndAddImageFromBase64(
+            ref.read(codiInsertProvider.notifier).pickAndAddItemImage(
                 item.photoPath,
                 item.itemName,
                 item.itemId,
